feat(swap): Setup useSwap Hook
