+++ app/src/views/swap.tsx
@@ -5,17 +5,11 @@
 
 import React from 'react'
 
-// Options
-import { SwapAndSendOptions } from '../options/select-and-send-options'
-
 // Context
 import { useSwapContext } from '../context/swap.context'
 
 // Hooks
-import { useWalletState } from '../state/wallet'
-
-// Types
-import { BlockchainToken, QuoteOption, WalletAccount } from '../constants/types'
+import { useSwap } from '../hooks/useSwap'
 
 // Components
 import { StandardButton, FlipTokensButton } from '../components/buttons'
@@ -37,103 +31,42 @@ import AdvancedIcon from '../assets/advanced-icon.svg'
 import { Row, Text, IconButton } from '../components/shared.styles'
 
 export const Swap = () => {
-  // Context
-  const { getLocale, getSwapQuotes } = useSwapContext()
-
-  // Wallet State
+  // Hooks
+  const swap = useSwap()
   const {
-    state: { tokenBalances, tokenList }
-  } = useWalletState()
+    fromAmount,
+    toAmount,
+    fromToken,
+    toToken,
+    insufficientBalance,
+    isFetchingQuote,
+    quoteOptions,
+    selectedQuoteOption,
+    selectingFromOrTo,
+    fromTokenBalance,
+    fiatValue,
+    swapAndSendSelected,
+    toAnotherAddress,
+    userConfirmedAddress,
+    selectedSwapSendAccount,
+    selectedSwapAndSendOption,
+    getTokenBalance,
+    onSelectFromToken,
+    onSelectToToken,
+    onSelectQuoteOption,
+    onClickFlipSwapTokens,
+    setSelectingFromOrTo,
+    handleOnSetFromAmount,
+    handleOnSetToAmount,
+    handleOnSetToAnotherAddress,
+    onCheckUserConfirmedAddress,
+    onSetSelectedSwapAndSendOption,
+    setSelectedSwapSendAccount,
+    setSwapAndSendSelected
+  } = swap
 
-  // ToDo: Setup useSwap hook where all this kind of state will be handled.
-  const price = 1519.28
-  const rate = 0.026
-
-  // State
-  const [fromToken, setFromToken] = React.useState<
-    BlockchainToken | undefined
-  >()
-  const [toToken, setToToken] = React.useState<BlockchainToken | undefined>()
-  const [fromAmount, setFromAmount] = React.useState<string>('')
-  const [toAmount, setToAmount] = React.useState<string>('')
-  const [isFetchingQuote, setIsFetchingQuote] = React.useState<
-    boolean | undefined
-  >(undefined)
-  const [selectingFromOrTo, setSelectingFromOrTo] = React.useState<
-    'from' | 'to' | undefined
-  >(undefined)
-  const [quoteOptions, setQuoteOptions] = React.useState<QuoteOption[]>([])
-  const [selectedQuoteOption, setSelectedQuoteOption] = React.useState<
-    QuoteOption | undefined
-  >()
-  const [selectedSwapAndSendOption, setSelectedSwapAndSendOption] =
-    React.useState<string>(SwapAndSendOptions[0].name)
-  const [swapAndSendSelected, setSwapAndSendSelected] =
-    React.useState<boolean>(false)
-  const [toAnotherAddress, setToAnotherAddress] = React.useState<string>('')
-  const [userConfirmedAddress, setUserConfirmedAddress] =
-    React.useState<boolean>(false)
-  const [selectedSwapSendAccount, setSelectedSwapSendAccount] = React.useState<
-    WalletAccount | undefined
-  >(undefined)
-
-  // Update on render
-  if (fromToken === undefined && fromToken !== tokenList[0]) {
-    setFromToken(tokenList[0])
-  }
-
-  // Effects
-  React.useEffect(() => {
-    let ignore = false
-    if (toToken && fromToken && fromAmount !== '') {
-      setIsFetchingQuote(true)
-      getSwapQuotes(
-        fromToken.contractAddress,
-        fromAmount,
-        toToken.contractAddress
-      )
-        .then((result) => {
-          if (!ignore) {
-            setQuoteOptions(result)
-            setSelectedQuoteOption(result[0])
-            setToAmount((Number(fromAmount) / rate).toFixed(6).toString())
-            setIsFetchingQuote(false)
-          }
-        })
-        .catch((error) => console.log(error))
-      return () => {
-        ignore = true
-      }
-    }
-  }, [toToken, fromToken, fromAmount, getSwapQuotes])
-
-  // Methods
-  const handleOnSetFromAmount = React.useCallback(
-    (value: string) => {
-      setFromAmount(value)
-      if (toToken) {
-        const updatedValue = (Number(value) / rate).toFixed(6).toString()
-        setToAmount(updatedValue)
-      }
-    },
-    [toToken, rate]
-  )
-
-  const handleOnSetToAmount = React.useCallback((value: string) => {
-    setToAmount(value)
-  }, [])
-
-  const getTokenBalance = React.useCallback(
-    (token: BlockchainToken): string => {
-      return tokenBalances ? tokenBalances[token.contractAddress] ?? '0' : '0'
-    },
-    [tokenBalances]
-  )
-
-  const onClickFlipSwapTokens = React.useCallback(() => {
-    setFromToken(toToken)
-    setToToken(fromToken)
-  }, [toToken, fromToken])
+  // Context
+  const { getLocale } = useSwapContext()
 
   const onClickSettings = React.useCallback(() => {
     // Todo: Add logic here to open settings view
@@ -143,70 +76,6 @@ export const Swap = () => {
   const onClickReviewOrder = React.useCallback(() => {
     // Todo: Add logic here to review order
   }, [])
-
-  const hideSelectTokenModal = React.useCallback(() => {
-    setSelectingFromOrTo(undefined) // hide modal
-  }, [])
-
-  const onSelectToToken = React.useCallback(
-    (token: BlockchainToken) => {
-      setToToken(token)
-      hideSelectTokenModal()
-    },
-    [hideSelectTokenModal]
-  )
-
-  const onSelectFromToken = React.useCallback(
-    (token: BlockchainToken) => {
-      setFromToken(token)
-      hideSelectTokenModal()
-    },
-    [hideSelectTokenModal]
-  )
-
-  const onSelectQuoteOption = React.useCallback((option: QuoteOption) => {
-    setSelectedQuoteOption(option)
-  }, [])
-
-  const onSetSelectedSwapAndSendOption = React.useCallback((value: string) => {
-    if (value === 'to-account') {
-      setToAnotherAddress('')
-    }
-    setSelectedSwapAndSendOption(value)
-  }, [])
-
-  const handleOnSetToAnotherAddress = React.useCallback((value: string) => {
-    setToAnotherAddress(value)
-  }, [])
-
-  const onCheckUserConfirmedAddress = React.useCallback(
-    (id: string, checked: boolean) => {
-      setUserConfirmedAddress(checked)
-    },
-    [userConfirmedAddress]
-  )
-
-  // Memos
-  const fromTokenBalance: number = React.useMemo(() => {
-    if (fromToken) {
-      return Number(getTokenBalance(fromToken))
-    }
-    return 0
-  }, [fromToken, tokenBalances, getTokenBalance])
-
-  const fiatValue: string | undefined = React.useMemo(() => {
-    if (fromAmount && price) {
-      return (price * Number(fromAmount)).toString()
-    }
-    return undefined
-  }, [price, fromAmount])
-
-  const insufficientBalance: boolean = React.useMemo(() => {
-    if (fromAmount && fromTokenBalance !== undefined) {
-      return Number(fromAmount) > fromTokenBalance
-    }
-    return false
-  }, [fromTokenBalance, fromAmount])
 
   // render
   return (
@@ -280,7 +149,7 @@ export const Swap = () => {
       </SwapContainer>
       {selectingFromOrTo && (
         <SelectTokenModal
-          onClose={hideSelectTokenModal}
+          onClose={() => setSelectingFromOrTo(undefined)}
           onSelectToken={
             selectingFromOrTo === 'from' ? onSelectFromToken : onSelectToToken
           }
